Update validation to allow for empty context lists
When no strings are provided in the contexts during evaluation, the context's feature's type in the dataset is the generic Sequence type rather than the expected Sequence[string], thus it fails validation with a ValueError, despite being a valid dataset. For example, if retrieval uses an unrealistically high similarity score threshold, no contexts will be found, but the error message will be the unhelpful and irrelevant `ValueError: Dataset feature "contexts" should be of type Sequence[string], got <class 'datasets.features.features.Sequence'>`.